recommendation_model/requirements.txt
# requirements.txt
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
google-generativeai==0.8.3
python-dotenv==1.0.1
ipykernel
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
